+++ Model/Collector/StoreUrlCollector.php
@@ -11,6 +11,7 @@ namespace Hryvinskyi\Csp\Model\Collector;
 
 use Hryvinskyi\Csp\Api\ConfigInterface;
 use Magento\Csp\Api\PolicyCollectorInterface;
+use Magento\Csp\Model\Collector\MergerInterface;
 use Magento\Csp\Model\Policy\FetchPolicy;
 use Magento\Framework\Url\ScopeResolverInterface;
 use Magento\Framework\UrlInterface;
@@ -22,6 +23,7 @@ class StoreUrlCollector implements PolicyCollectorInterface
     public function __construct(
         private readonly ScopeResolverInterface $scopeResolver,
         private readonly ConfigInterface $config,
+        private readonly MergerInterface $merger
     ) {
     }
 
@@ -41,7 +43,16 @@ class StoreUrlCollector implements PolicyCollectorInterface
         $storeUrls = $this->getStoreUrls();
 
         foreach (FetchPolicy::POLICIES as $directive) {
-            $policies[] = $this->createFetchPolicy($directive, $storeUrls);
+            $policy = $this->createFetchPolicy($directive, $storeUrls);
+            if (array_key_exists($directive, $policies)) {
+                if ($this->merger->canMerge($policies[$directive], $policy)) {
+                    $policies[$directive] = $this->merger->merge($policies[$directive], $policy);
+                } else {
+                    throw new \RuntimeException('Cannot merge a policy of ' . get_class($policy));
+                }
+            } else {
+                $policies[$directive] = $policy;
+            }
         }
 
         return $policies;